Update docs / move story/name utils into tests folder
