+++ packages/nimble-components/README.md
@@ -98,9 +98,8 @@ The goal of the Nimble design system is to provide a consistent style for applic
 
 ## Localization
 
-Some Nimble components use strings/labels that need to be localized, if the consuming application supports localization. Nimble exposes these localizable labels as design tokens, to support both localization and the ability for clients to override the strings.
+Most user-visible strings displayed by Nimble components are provided by the client application and are expected to be localized by the application if necessary. However, some strings are built into Nimble components and are provided only in English. An application can provide localized versions of these strings by using design tokens set on label provider elements.
 
-Nimble provides English strings as the token defaults, and provides `nimble-label-provider-*` elements with APIs for overriding those values.  
 There are currently 2 label providers:
 
 -   `nimble-label-provider-core`: Used for labels for all components besides the table
