+++ packages/nimble-components/src/label-provider/table/tests/label-provider-table.stories.ts
@@ -2,7 +2,7 @@ import type { StoryObj } from '@storybook/html';
 import {
     LabelProviderArgs,
     labelProviderMetadata
-} from '../../base/label-provider-stories-utils';
+} from '../../base/tests/label-provider-stories-utils';
 import { labelProviderTableTag } from '..';
 import { removeTablePrefixAndCamelCase } from '../name-utils';
 import * as labelTokensNamespace from '../label-tokens';
