+++ packages/nimble-components/src/table/tests/table.stories.ts
@@ -15,7 +15,7 @@ import { tableColumnTextTag } from '../../table-column/text';
 import {
     addLabelUseMetadata,
     type LabelUserArgs
-} from '../../label-provider/base/label-user-stories-utils';
+} from '../../label-provider/base/tests/label-user-stories-utils';
 import { labelProviderTableTag } from '../../label-provider/table';
 
 interface TableArgs extends LabelUserArgs {
